Epic2 - Mariia Stepova

+++ ai_12/mariia_stepova/епік2/1.5.cpp
@@ -3,9 +3,9 @@
 
 int main()
 {
-    float a = 1000;
-    float b = 0.0001;
-    float x = pow(a+b,2) , y = pow(a, 2) + 2*a*b , z = pow(b, 2) , R;
+    double a = 1000;
+    double b = 0.0001;
+    double x = pow(a+b,2) , y = pow(a, 2) + 2*a*b , z = pow(b, 2) , R;
 
     R = (x-y)/z;
